REC_1:rename packages + add new dependencies + refacto hibernate session factory build + add unitary test AgentDaoTest
