+++ reclamation-model/src/main/java/com/sjl/reclamation/model/dao/helpers/HibernateConfiguration.java
@@ -1,0 +1,27 @@
+package com.sjl.reclamation.model.dao.helpers;
+
+import org.hibernate.SessionFactory;
+import org.hibernate.cfg.Configuration;
+
+public class HibernateConfiguration {
+
+    private static SessionFactory sessionFactory;
+
+    // Method Used to create the hibernate's SessionFactory Object
+    private static SessionFactory buildSessionFactory() {
+        return new Configuration().configure().buildSessionFactory();
+    }
+
+    /**
+     * Singleton factory
+     * @return @{@link SessionFactory} instance
+     */
+    public synchronized  static SessionFactory sessionFactory(){
+        if(sessionFactory==null)
+            sessionFactory = buildSessionFactory();
+        return sessionFactory;
+    }
+
+
+
+}
