--- reclamation-model/src/main/java/com/zineb/hibernate/dao/DaoInterface.java
@@ -1,15 +1,0 @@
-package com.zineb.hibernate.dao;
-
-import java.io.Serializable;
-import java.util.List;
-
-public interface DaoInterface<T , Id extends Serializable >{
-	
-	public void persist(T entity);
-	public void update (T entity);
-	public T findbyId (Id id);
-	public void delete(T entity);
-	public List<T> findAll();
-	public void deleteAll();
-
-}
