--- reclamation-model/src/main/java/com/zineb/hibernate/entity/Camion.java
@@ -1,49 +1,0 @@
-package com.zineb.hibernate.entity;
-
-import java.io.Serializable;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Entity
-public class Camion implements Serializable{
-	
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-
-
-	@Id @GeneratedValue
-	@Column (name="id")
-	private Long id;
-	
-
-	@Column (name="Matricule")
-	private String Matricule;
-	
-	
-	public Long getCamionId() {
-		return id;
-	}
-
-	public void setCamionId(Long id) {
-		this.id = id;
-	}
-
-	public String getCamionMatricule() {
-		return Matricule;
-	}
-
-	public void setCamionMatricule(String matricule) {
-		Matricule = matricule;
-	}
-	
-	public Camion(){
-		
-	}
-
-
-}
