+++ reclamation-model/src/main/resources/hibernate.cfg.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!DOCTYPE hibernate-configuration PUBLIC
-"-//Hibernate/Hibernate Configuration DTD 3.0//EN"
-"http://hibernate.sourceforge.net/hibernate-configuration-3.0.dtd">
+        "-//Hibernate/Hibernate Configuration DTD 3.0//EN"
+        "http://hibernate.sourceforge.net/hibernate-configuration-3.0.dtd">
 <hibernate-configuration>
- 	<session-factory>
-  		<property name="hibernate.connection.driver_class">com.mysql.jdbc.Driver</property>
-  		<property name="hibernate.connection.url">jdbc:mysql://localhost/Test</property>
-  		<property name="hibernate.connection.username">root</property>
-  		<property name="hibernate.connection.password">ziben</property>
-  		<property name="hibernate.dialect">org.hibernate.dialect.MySQLDialect</property>
-  		<property name="hibernate.default_schema">MKYONG</property>
-  		<property name="show_sql">true</property>
-  		<mapping class="com.zineb.hibernate.entity.Agent"/>
-  		<mapping class="com.zineb.hibernate.entity.Camion"/>
-  
-	</session-factory>
+    <session-factory>
+        <property name="hibernate.connection.driver_class">com.mysql.jdbc.Driver</property>
+        <property name="hibernate.connection.url">jdbc:mysql://localhost/Test</property>
+        <property name="hibernate.connection.username">root</property>
+        <property name="hibernate.connection.password">ziben</property>
+        <property name="hibernate.dialect">org.hibernate.dialect.MySQLDialect</property>
+        <property name="hibernate.default_schema">MKYONG</property>
+        <property name="show_sql">true</property>
+        <mapping class="com.sjl.reclamation.model.entity.Agent"/>
+        <mapping class="com.sjl.reclamation.model.entity.Camion"/>
+
+    </session-factory>
 </hibernate-configuration>
