+++ reclamation-model/src/test/resources/log4j.properties
@@ -1,0 +1,12 @@
+log4j.appender.stdout=org.apache.log4j.ConsoleAppender
+log4j.appender.stdout.Target=System.out
+log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
+log4j.appender.stdout.layout.ConversionPattern=%d{HH:mm:ss,SSS} %-5p [%c] - %m%n
+
+log4j.rootLogger=info, stdout
+# basic log level for all messages
+log4j.logger.org.hibernate=info
+
+# SQL statements and parameters
+log4j.logger.org.hibernate.SQL=info
+log4j.logger.org.hibernate.type.descriptor.sql=info
