+++ reclamation-rest-api/pom.xml
@@ -1,36 +1,36 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
         <artifactId>sjl_reclamation</artifactId>
-        <groupId>com.zineb</groupId>
+        <groupId>com.sjl.reclamation</groupId>
         <version>0.0.1-SNAPSHOT</version>
     </parent>
 
-  
-  <artifactId>reclamation-rest-api</artifactId>
-  <packaging>war</packaging>
 
-  <name>reclamation-rest-api Maven Webapp</name>
-  <!-- FIXME change it to the project's website -->
-  <url>http://www.example.com</url>
+    <artifactId>reclamation-rest-api</artifactId>
+    <packaging>war</packaging>
 
-  
+    <name>reclamation-rest-api Maven Webapp</name>
+    <!-- FIXME change it to the project's website -->
+    <url>http://www.example.com</url>
 
-  <dependencies>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-    <groupId>com.zineb</groupId>
-    <artifactId>reclamation-service</artifactId>
-    </dependency>
-  </dependencies>
 
-  <build>
-    <finalName>reclamation-rest-api</finalName>
-    
-  </build>
+    <dependencies>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.zineb</groupId>
+            <artifactId>reclamation-service</artifactId>
+        </dependency>
+    </dependencies>
+
+    <build>
+        <finalName>reclamation-rest-api</finalName>
+
+    </build>
 </project>
